+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `Wildfire` will be documented in this file.
 
+## [Unreleased](https://github.com/rougin/wildfire/compare/v0.2.1...HEAD)
+
+### Fixed
+- Issue in getting foreign tables
+
 ## [0.2.1](https://github.com/rougin/wildfire/compare/v0.2.0...v0.2.1) - 2016-05-14
 
 ### Changed